Inverse the dependencies
