+++ addon/components/em-form-group.js
@@ -159,13 +159,19 @@ export default Ember.Component.extend(HasPropertyMixin, HasPropertyValidationMix
     }
   }),
 
-  propertyOptions: Ember.computed('property', 'validations.attrs.@each.options', function() {
+  required: Ember.computed('property', 'validations.attrs.@each.options.presence.presence', function () {
     const property = this.get('property');
 
-    return this.get(`model.validations.attrs.${property}.options`) || false;
+    return this.get(`model.validations.attrs.${property}.options.presence.presence`) || false;
   }),
 
-  required: Ember.computed('propertyOptions.presence.presence', function () {
-    return this.get('propertyOptions.presence.presence') || false;
+  id: Ember.computed('inputId', 'cid', 'property', {
+    get: function() {
+      if (this.get('cid')) {
+        return this.get('cid');
+      } else {
+        return `${this.get('property')}-${this.elementId}`;
+      }
+    }
   })
 });
