+++ addon/components/em-input.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import FormGroupComponent from './em-form-group';
+import layout from '../templates/components/em-input';
 
 /*
 Form Input
@@ -7,7 +7,8 @@ Form Input
 Syntax:
 {{em-input property="property name"}}
  */
-export default FormGroupComponent.extend({
+export default Ember.Component.extend({
+  layout: layout,
   elementClass: null,
   htmlComponent: 'erf-html-input',
   property: null,
@@ -19,12 +20,57 @@ export default FormGroupComponent.extend({
   readonly: null,
   autoresize: null,
   disabled: null,
+  canShowErrors: false,
   controlWrapper: Ember.computed('form.formLayout', {
     get: function() {
       if (this.get('form.formLayout') === 'horizontal') {
         return 'col-sm-10';
       }
       return null;
+    }
+  }),
+  didReceiveAttrs( /*attrs*/ ) {
+    this._super(...arguments);
+    // set it to the correct value of the selection
+    this.selectedValue = Ember.computed.alias('model.' + this.get('property'));
+  },
+
+  /*
+  Observes the helpHasErrors of the help control and modify the 'status' property accordingly.
+   */
+  focusIn() {
+    if (this.get('form.showErrorsOnFocusIn')) {
+      return this.set('canShowErrors', true);
+    }
+  },
+
+  /*
+  Listen to the focus out of the form group and display the errors
+   */
+  focusOut() {
+    return this.set('canShowErrors', true);
+  },
+
+  /*
+  Listen to the keyUp of the form group and display the errors if showOnKeyUp is true.
+   */
+  keyUp() {
+    if (this.get('showOnKeyUp')) {
+      return this.set('canShowErrors', true);
+    }
+  },
+
+  hideValidationsOnFormChange: Ember.observer('form', 'form.model', function() {
+    this.set('canShowErrors', false);
+  }),
+
+  id: Ember.computed('cid', 'property', {
+    get: function() {
+      if (this.get('cid')) {
+        return this.get('cid');
+      } else {
+        return `${this.get('property')}-${this.elementId}`;
+      }
     }
   })
 });
