--- addon/components/html-input.js
@@ -1,11 +1,0 @@
-import Ember from 'ember';
-import layout from '../templates/components/html-input';
-
-export default Ember.Component.extend({
-  layout: layout,
-  didReceiveAttrs( /*attrs*/ ) {
-    this._super(...arguments);
-    // set it to the correct value of the selection
-    this.selectedValue = Ember.computed.alias('mainComponent.model.' + this.get('mainComponent.property'));
-  }
-});
