+++ addon/templates/components/em-checkbox.hbs
@@ -1,0 +1,21 @@
+{{#em-form-group model=model property=property label=label form=form
+  label=label yieldInLabel=yieldInLabel
+  labelWrapperClass=labelWrapperClass labelClass=labelClass
+  help=help controlWrapper=controlWrapper
+  cid=cid validationIcons=validationIcons canShowErrors=canShowErrors inputComponent=this}}
+  {{erf-html-checkbox
+
+    placeholder=placeholder
+    value=selectedValue
+    name=name
+    disabled=disabled
+    class=(concat "form-control " elementClass)
+    id=cid
+    required=required
+    title=title
+    pattern=pattern
+    autofocus=autofocus
+    readonly=readonly
+    autoresize=autoresize
+  }}
+{{/em-form-group}}
