+++ addon/templates/components/form-group-control.hbs
@@ -1,11 +1,15 @@
-{{#if controlWrapper}}
-    <div class={{controlWrapper}}>
-        {{#component mainComponent.htmlComponent mainComponent=mainComponent form=form}}
-            {{yield}}
-        {{/component}}
-    </div>
+{{#if (eq mainComponent.htmlComponent 'erf-html-input')}}
+  {{yield}}
 {{else}}
-    {{#component mainComponent.htmlComponent mainComponent=mainComponent form=form}}
-        {{yield}}
-    {{/component}}
+  {{#if controlWrapper}}
+      <div class={{controlWrapper}}>
+          {{#component mainComponent.htmlComponent mainComponent=mainComponent form=form}}
+              {{yield}}
+          {{/component}}
+      </div>
+  {{else}}
+      {{#component mainComponent.htmlComponent mainComponent=mainComponent form=form}}
+          {{yield}}
+      {{/component}}
+  {{/if}}
 {{/if}}
