--- addon/templates/components/html-input.hbs
@@ -1,16 +1,0 @@
-{{input
-  type=mainComponent.type
-
-  placeholder=mainComponent.placeholder
-  value=selectedValue
-  name=mainComponent.name
-  disabled=mainComponent.disabled
-  class=(concat "form-control " mainComponent.elementClass)
-  id=mainComponent.id
-  required=mainComponent.required
-  title=mainComponent.title
-  pattern=mainComponent.pattern
-  autofocus=mainComponent.autofocus
-  readonly=mainComponent.readonly
-  autoresize=mainComponent.autoresize
-}}
