--- app/components/erf-html-input.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-rapid-forms/components/html-input';
